# Improve doublylikedlist IndexOf() by reducing the space complexity from O(n) to O(1)