--- a/lists/doublylinkedlist/doublylinkedlist.go
+++ b/lists/doublylinkedlist/doublylinkedlist.go
@@ -88,7 +88,7 @@ func (list *List[T]) Get(index int) (T, bool) {
 		return t, false
 	}
 
-	// determine traveral direction, last to first or first to last
+	// determine traversal direction, last to first or first to last
 	if list.size-index < index {
 		element := list.last
 		for e := list.size - 1; e != index; e, element = e-1, element.prev {
@@ -172,7 +172,7 @@ func (list *List[T]) Contains(values ...T) bool {
 
 // Values returns all elements in the list.
 func (list *List[T]) Values() []T {
-	values := make([]T, list.size, list.size)
+	values := make([]T, list.size)
 	for e, element := 0, list.first; element != nil; e, element = e+1, element.next {
 		values[e] = element.value
 	}
@@ -181,12 +181,9 @@ func (list *List[T]) Values() []T {
 
 // IndexOf returns index of provided element
 func (list *List[T]) IndexOf(value T) int {
-	if list.size == 0 {
-		return -1
-	}
-	for index, element := range list.Values() {
-		if element == value {
-			return index
+	for e, element := 0, list.first; element != nil; e, element = e+1, element.next {
+		if element.value == value {
+			return e
 		}
 	}
 	return -1