@ksw2000 2024-06-21
There is the same issue in `singlylinkedlist`